ESYS: Add esys_free helper function
Add esys_free, a wrapper around free(), to ESYS when building on Windows.
Having a wrapper around free is best practice when building libraries
on Windows; this is because there is a high chance that the caller
program will be built using a different version of the C runtime
than the library, which means that the program and the library will have
separate heaps. This is an issue because many of the ESYS functions
allocate memory for the caller and expect the caller to free the memory,
but in the case where the caller was built with a different version of
the C runtime, the caller will not be able to free the memory because
it does not have access to the esys heap. esys_free allows the caller
program to have the library free any memory that the caller didn't allocate.

Signed-off-by: David Maria <davidjmaria@fb.com>